pyproject.toml
[tool.poetry]
name = "szwch"
version = "0.1.0"
description = ""
authors = ["Kamil <kamil@patientsource.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
uvloop = "^0.17.0"
uvicorn = "^0.21.1"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
pytest = "^7.2.2"
hiredis = "^2.2.2"
redis = "^4.5.3"
psycopg2 = "^2.9.5"
whitenoise = "^6.4.0"
hashids = "^1.3.1"
drf-spectacular = "^0.26.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
